+++ plugin/src/main/java/de/infolektuell/gradle/jextract/extensions/InstallationHandler.java
@@ -1,6 +1,7 @@
 package de.infolektuell.gradle.jextract.extensions;
 
 import org.gradle.api.file.DirectoryProperty;
+import org.gradle.api.file.RegularFileProperty;
 import org.gradle.api.provider.Property;
 import org.gradle.jvm.toolchain.JavaLanguageVersion;
 import org.jspecify.annotations.NonNull;
@@ -16,5 +17,10 @@ public abstract class InstallationHandler {
     /// The path of a local Jextract installation.
     /// If it is set, this installation is used by the plugin.
     /// @return A property to configure the JExtract installation path
+    /// @deprecated Use Gradle property `org.openjdk.jextract.installaption-path` instead.
     public abstract DirectoryProperty getLocation();
+
+    /// A [java.util.Properties] file containing the remote locations where to download the Jextract distributions
+    /// @return a property
+    public abstract RegularFileProperty getDistributions();
 }